RenderText’s minimum preferred width is incorrect automatic hyphenation is used https://bugs.webkit.org/show_bug.cgi?id=89814
Reviewed by Sam Weinig.

Source/WebCore:

Test: fast/text/hyphen-min-preferred-width.html

* rendering/RenderText.cpp:
(WebCore::maxWordFragmentWidth): Added this helper function. It returns the width of the
widest hyphenated fragment of the word (except for the suffix after the last hyphen) and
the offset of the last hyphenation opportunity.
(WebCore::RenderText::computePreferredLogicalWidths): Changed this function to consider the
widths of hyphenated word fragments rather than entire words when computing the minimum
width. Because hyphenating and measuring all fragments is expensive, hyphenation is only
attempted if the unhyphenated word is wider than the current minimum width.

LayoutTests:

* fast/text/hyphen-min-preferred-width-expected.html: Added.
* fast/text/hyphen-min-preferred-width.html: Added.



Identifier: 107680@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121102 268f45cc-cd09-0410-ab3c-d52691b4dbfc