[Chromium] Release media resources after each LayoutTest on Android https://bugs.webkit.org/show_bug.cgi?id=89720
Reviewed by Eric Carlson.

The LayoutTests can hang when run on Android because we sometimes run
out of media resources. This patch causes us to release our media
resources after each test in order to avoid running out of this
resource. In production, there are other mechanisms that manage this
resource.

* DumpRenderTree/chromium/TestShell.cpp:
(TestShell::resetTestController):


Identifier: 107695@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121118 268f45cc-cd09-0410-ab3c-d52691b4dbfc