[BlackBerry] Use gesture SwipeDown to exit fullscreen for both video and plugin. https://bugs.webkit.org/show_bug.cgi?id=89960
Reviewed by Antonio Gomes.

We used to use gesture swipedown to exit fullscreen for plugin, but not fullscreen
HTML5 video; When a swipe down happens, it applies this event to all the pluginviews
in a page, even though only the one in fullscreen mode will process this.

With this patch, the SwipeDown gesture will only apply to the fullscreen elemement,
which is either a plugin, or an Video element.

* Api/WebPage.cpp:
(BlackBerry::WebKit::WebPage::notifySwipeEvent):


Identifier: 107815@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121245 268f45cc-cd09-0410-ab3c-d52691b4dbfc