REGRESSION (Safari 5?): Pasting a line into textarea inserts two newlines https://bugs.webkit.org/show_bug.cgi?id=49288
Reviewed by Tony Chang.

Source/WebCore:

The bug was caused by positionAvoidingPrecedingNodes getting out of a block when the insertion point is at a line break.
It caused the subsequent code to be misinformed of the insertion position and ended up not pruning the extra line break.

Fixed the bug by checking this special case and bailing out so that we don't crawl out of the enclosing block.
It's similar to checks several lines below it.

Test: editing/pasteboard/copy-paste-pre-line-content.html

* editing/ReplaceSelectionCommand.cpp:
(WebCore::positionAvoidingPrecedingNodes):

LayoutTests:

Add a test regerssion test for copying & pasting a line in pre into textarea twice.

* editing/pasteboard/copy-paste-pre-line-content-expected.txt: Added.
* editing/pasteboard/copy-paste-pre-line-content.html: Added.


Identifier: 107926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121357 268f45cc-cd09-0410-ab3c-d52691b4dbfc