[gtk] Spell checker doesn't recognize contractions (apostrophes) https://bugs.webkit.org/show_bug.cgi?id=86118
Patch by Martin Robinson <mrobinson@igalia.com> on 2012-06-27
Reviewed by Gustavo Noronha Silva.

Work-around a bug in Pango by trying to detect apostrophes
that create contractions. This work-around is similar to one
found in gtkspell.

* webkit/webkitspellcheckerenchant.cpp:
(wordEndIsAContractionApostrophe): Added this helper which tries to detect
situations where a word end is both an apostrophe and followed by a alphabetic
character.
(checkSpellingOfString): When searching for the end of a word, skip over
apostrophes that appear to be part of contractions.

Identifier: 107929@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121360 268f45cc-cd09-0410-ab3c-d52691b4dbfc