[Chromium] WebFontRendering.cpp requires some OS(ANDROID) ifdefs to build downstream https://bugs.webkit.org/show_bug.cgi?id=90292
Reviewed by Nate Chapin.

These ifdefs are required to build this file downstream. There's some
sublte difference between how the OS flags are set upstream and
downstream. This is on our list of issues to resolve, but in the
meantime, this patch makes these files identical upstream and
downstream to reduce noise in the upstreaming queue.

* src/linux/WebFontRendering.cpp:
(WebKit::WebFontRendering::setSubpixelPositioning):


Identifier: 108159@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121594 268f45cc-cd09-0410-ab3c-d52691b4dbfc