garden-o-matic broken: TypeError: 'undefined' is not an object (evaluating 'buildLocations[currentIndex].url') https://bugs.webkit.org/show_bug.cgi?id=90243
Reviewed by Dirk Pranke.

jQuery was trying to be too smart and parsing the jsonp as json because of it's content-type.
Excise jQuery and just use XHR directly since it's easier to maintain something where we control it
all.

Not really sure how to unittest this. I tested it all manually of course.

* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/net.js:
Made net.ajax a drop-in replacement for the features of $.ajax that we were using.
* BuildSlaveSupport/build.webkit.org-config/public_html/TestFailures/scripts/ui_unittests.js:
Not related to this patch, but figured I'd update the failing test while I was at it.


Identifier: 108182@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121617 268f45cc-cd09-0410-ab3c-d52691b4dbfc