Make the skia_test_expectations.txt file optional. https://bugs.webkit.org/show_bug.cgi?id=90400
Reviewed by Dirk Pranke.

It used to be optional. This regressed at some point. It's important that it be
optional so that webkit-patch commands work in a pure-webkit checkout for chromium bots.
Specifically, this was breaking webkit-patch rebaseline-test when it would go to update
TestExpectations.

* Scripts/webkitpy/layout_tests/port/chromium.py:
(ChromiumPort.expectations_files):
* Scripts/webkitpy/layout_tests/port/chromium_unittest.py:
(ChromiumDriverTest.test_expectations_dict):

Identifier: 108344@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121784 268f45cc-cd09-0410-ab3c-d52691b4dbfc